:recycle: refactor: QType에 domain 패키지로 도메인 코드 이동 반영
