+++ src/main/java/tipitapi/drawmytoday/domain/admin/controller/AdminController.java
@@ -17,11 +17,11 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import tipitapi.drawmytoday.domain.admin.dto.GetDiaryAdminResponse;
-import tipitapi.drawmytoday.domain.admin.service.AdminService;
 import tipitapi.drawmytoday.common.resolver.AuthUser;
 import tipitapi.drawmytoday.common.response.SuccessResponse;
 import tipitapi.drawmytoday.common.security.jwt.JwtTokenInfo;
+import tipitapi.drawmytoday.domain.admin.dto.GetDiaryAdminResponse;
+import tipitapi.drawmytoday.domain.admin.service.AdminService;
 
 @RestController
 @RequestMapping("/admin")
