+++ src/main/java/tipitapi/drawmytoday/domain/adreward/service/AdRewardService.java
@@ -5,9 +5,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import tipitapi.drawmytoday.domain.adreward.domain.AdReward;
 import tipitapi.drawmytoday.domain.adreward.repository.AdRewardRepository;
-import tipitapi.drawmytoday.domain.user.service.ValidateUserService;
 import tipitapi.drawmytoday.domain.ticket.service.TicketService;
 import tipitapi.drawmytoday.domain.user.domain.User;
+import tipitapi.drawmytoday.domain.user.service.ValidateUserService;
 
 @Service
 @Transactional(readOnly = true)
