+++ src/main/java/tipitapi/drawmytoday/domain/user/service/UserService.java
@@ -3,9 +3,9 @@ package tipitapi.drawmytoday.domain.user.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import tipitapi.drawmytoday.domain.user.repository.UserRepository;
 import tipitapi.drawmytoday.domain.user.domain.SocialCode;
 import tipitapi.drawmytoday.domain.user.domain.User;
+import tipitapi.drawmytoday.domain.user.repository.UserRepository;
 
 @Service
 @Transactional(readOnly = true)
