+++ src/main/java/tipitapi/drawmytoday/domain/user/service/ValidateUserService.java
@@ -3,9 +3,9 @@ package tipitapi.drawmytoday.domain.user.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import tipitapi.drawmytoday.domain.user.exception.UserAccessDeniedException;
 import tipitapi.drawmytoday.domain.user.domain.SocialCode;
 import tipitapi.drawmytoday.domain.user.domain.User;
+import tipitapi.drawmytoday.domain.user.exception.UserAccessDeniedException;
 import tipitapi.drawmytoday.domain.user.exception.UserNotFoundException;
 import tipitapi.drawmytoday.domain.user.repository.UserRepository;
